transforms/language/similarity/tools/load_elasticsearch/requirements.txt

python-dotenv>=1.0.1
requests
sentence-transformers>=3.3.1